requirements.txt
attrs==19.3.0
awscli==1.15.84
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
cffi==1.13.1
cycler==0.10.0
Cython==0.29.13
decorator==4.4.1
defusedxml==0.6.0
dill==0.2.7.1
entrypoints==0.3
graphviz==0.13.2
h5py==2.10.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.3
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
ptyprocess==0.6.0
pyyaml==3.13
pyasn1==0.4.4
pycocotools==2.0.0
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.5
python-dateutil==2.8.0
pytz==2019.3
pyzmq==18.1.0
qtconsole==4.5.5
rsa==3.4.2
Send2Trash==1.5.0
six==1.12.0
tensorboardX==1.9
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0

